[ALPS07927427] Fix Coverity of untrusted loop boundary
The length of DHCP option fields is calculated from DHCP
datagram size, and Coverity considered it as tainted value.

This patch adds a comparison and update loop bound counter
to ensure the tainted values are properly sanitized,
by checking that the value is within a permissible range.

MTK-Commit-Id: 8c56ff47b590a185e850fd97f64cb3b135202d78

CR-Id: ALPS07927427
Feature: Wi-Fi Driver CONNAC
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/7336743
CheckForbiddenKeyword: srv_pis <srv_pis@mediatek.com>
Test: srv_mspautosanity
Coverity-Review-Label: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Test: srv_pf_nep_sanity
CheckPatch: srv_pis <srv_pis@mediatek.com>
Build: srv_pf_nep_win <srv_pf_nep_win@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
CheckCRInfo: srv_pis <srv_pis@mediatek.com>
Coverity-Review-Label: srv_pf_nep_win <srv_pf_nep_win@mediatek.com>
CheckCourse: srv_pis <srv_pis@mediatek.com>
Build: srv_preflight_b001 <srv_preflight_b001@mediatek.com>
Test: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
CheckSWIP: srv_pis <srv_pis@mediatek.com>

+++ include/nic/que_mgt.h
@@ -1290,7 +1290,7 @@ void qmArpMonitorHandlePkt(struct ADAPTER *prAdapter,
 #endif /* CFG_QM_ARP_MONITOR_MSG */
 uint8_t *qmGetArpPkt(uint8_t *pucData, uint16_t u2PacketLen);
 struct DHCP_PROTOCOL *qmGetDhcpPkt(uint8_t *pucData, uint16_t u2PacketLen,
-	u_int8_t fgFromServer, uint32_t *pDhcpLen);
+	u_int8_t fgFromServer, uint16_t *pDhcpLen);
 #endif
 
 #if defined(CFG_SUPPORT_REPLAY_DETECTION) || \
